move things to a separate dir
